Reorganise GovSpec, add GovCertSpec
