+++ eras/conway/impl/cardano-ledger-conway.cabal
@@ -118,6 +118,7 @@ library testlib
         Test.Cardano.Ledger.Conway.Imp.EpochSpec
         Test.Cardano.Ledger.Conway.Imp.EnactSpec
         Test.Cardano.Ledger.Conway.Imp.GovSpec
+        Test.Cardano.Ledger.Conway.Imp.GovCertSpec
         Test.Cardano.Ledger.Conway.Imp.UtxoSpec
         Test.Cardano.Ledger.Conway.Imp.UtxosSpec
         Test.Cardano.Ledger.Conway.Imp.RatifySpec
