+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/EnactSpec.hs
@@ -4,20 +4,22 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
 
 module Test.Cardano.Ledger.Conway.Imp.EnactSpec (spec) where
 
 import Cardano.Ledger.Address
-import Cardano.Ledger.BaseTypes (EpochInterval (..), ProtVer (..), StrictMaybe (..), succVersion)
+import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Coin
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Governance
-import Cardano.Ledger.Conway.Rules (EnactSignal (..))
-import Cardano.Ledger.Credential (Credential (..))
-import Cardano.Ledger.Shelley.LedgerState (
-  curPParamsEpochStateL,
-  nesEsL,
- )
+import Cardano.Ledger.Conway.PParams
+import Cardano.Ledger.Conway.Rules
+import Cardano.Ledger.Credential
+import Cardano.Ledger.Shelley.LedgerState
+import Data.Default.Class (def)
+import qualified Data.Map.Strict as Map
+import qualified Data.Sequence as Seq
 import Lens.Micro
 import Test.Cardano.Ledger.Conway.ImpTest
 import Test.Cardano.Ledger.Core.Rational
@@ -25,7 +27,9 @@ import Test.Cardano.Ledger.Imp.Common
 
 spec ::
   forall era.
-  ConwayEraImp era =>
+  ( ConwayEraImp era
+  , GovState era ~ ConwayGovState era
+  ) =>
   SpecWith (ImpTestState era)
 spec =
   describe "ENACT" $ do
@@ -106,3 +110,107 @@ spec =
       submitYesVote_ (StakePoolVoter stakePoolId2) govActionId
       passNEpochs 2
       getsNES (nesEsL . curPParamsEpochStateL . ppProtocolVersionL) `shouldReturn` nextProtVer
+    it "NoConfidence" $ do
+      modifyPParams $ \pp ->
+        pp
+          & ppDRepVotingThresholdsL . dvtCommitteeNoConfidenceL .~ 1 %! 2
+          & ppPoolVotingThresholdsL . pvtCommitteeNoConfidenceL .~ 1 %! 2
+          & ppCommitteeMaxTermLengthL .~ EpochInterval 200
+      let
+        getCommittee =
+          getsNES $
+            nesEsL . esLStateL . lsUTxOStateL . utxosGovStateL . committeeGovStateL
+        assertNoCommittee :: HasCallStack => ImpTestM era ()
+        assertNoCommittee =
+          do
+            committee <- getCommittee
+            impAnn "There should not be a committee" $ committee `shouldBe` SNothing
+      assertNoCommittee
+      khCC <- freshKeyHash
+      (drep, _, _) <- setupSingleDRep 1_000_000
+      let committeeMap =
+            Map.fromList
+              [ (KeyHashObj khCC, EpochNo 50)
+              ]
+      prevGaidCommittee@(GovPurposeId gaidCommittee) <-
+        electCommittee
+          SNothing
+          drep
+          mempty
+          committeeMap
+      (khSPO, _, _) <- setupPoolWithStake $ Coin 42_000_000
+      logStakeDistr
+      submitYesVote_ (StakePoolVoter khSPO) gaidCommittee
+      replicateM_ 4 passEpoch
+      impAnn "Committee should be elected" $ do
+        committee <- getCommittee
+        committee `shouldBe` SJust (Committee committeeMap $ 1 %! 2)
+      pp <- getsNES $ nesEsL . curPParamsEpochStateL
+      returnAddr <- registerRewardAccount
+      gaidNoConf <-
+        submitProposal $
+          ProposalProcedure
+            { pProcReturnAddr = returnAddr
+            , pProcGovAction = NoConfidence (SJust prevGaidCommittee)
+            , pProcDeposit = pp ^. ppGovActionDepositL
+            , pProcAnchor = def
+            }
+      submitYesVote_ (StakePoolVoter khSPO) gaidNoConf
+      submitYesVote_ (DRepVoter $ KeyHashObj drep) gaidNoConf
+      replicateM_ 4 passEpoch
+      assertNoCommittee
+    it "Constitution" $ do
+      (dRep, committeeMember, _) <- electBasicCommittee
+      (govActionId, constitution) <- submitConstitution SNothing
+
+      ConwayGovState proposalsBeforeVotes enactStateBeforeVotes pulserBeforeVotes _ _ _ <-
+        getsNES $ newEpochStateGovStateL @era
+      submitYesVote_ (DRepVoter dRep) govActionId
+      submitYesVote_ (CommitteeVoter committeeMember) govActionId
+      ConwayGovState proposalsAfterVotes enactStateAfterVotes pulserAfterVotes _ _ _ <-
+        getsNES newEpochStateGovStateL
+
+      impAnn "Votes are recorded in the proposals" $ do
+        let proposalsWithVotes =
+              proposalsAddVote
+                (CommitteeVoter committeeMember)
+                VoteYes
+                govActionId
+                ( proposalsAddVote
+                    (DRepVoter dRep)
+                    VoteYes
+                    govActionId
+                    proposalsBeforeVotes
+                )
+        proposalsAfterVotes `shouldBe` proposalsWithVotes
+
+      impAnn "EnactState has not changed" $
+        enactStateAfterVotes `shouldBe` enactStateBeforeVotes
+
+      impAnn "Pulser has not changed" $
+        pulserAfterVotes `shouldBe` pulserBeforeVotes
+
+      passEpoch
+
+      impAnn "New constitution is not enacted after one epoch" $ do
+        constitutionAfterOneEpoch <- getsNES $ newEpochStateGovStateL . constitutionGovStateL
+        constitutionAfterOneEpoch `shouldBe` def
+
+      impAnn "Pulser should reflect the constitution to be enacted" $ do
+        ConwayGovState _ _ _ _ _ pulser <- getsNES newEpochStateGovStateL
+        let ratifyState = extractDRepPulsingState pulser
+        gasId <$> rsEnacted ratifyState `shouldBe` govActionId Seq.:<| Seq.Empty
+        rsEnactState ratifyState ^. ensConstitutionL `shouldBe` constitution
+
+      passEpoch
+
+      impAnn "Constitution is enacted after two epochs" $ do
+        curConstitution <- getsNES $ newEpochStateGovStateL . constitutionGovStateL
+        curConstitution `shouldBe` constitution
+
+      impAnn "Pulser is reset" $ do
+        ConwayGovState _ _ _ _ _ pulser <- getsNES newEpochStateGovStateL
+        let pulserRatifyState = extractDRepPulsingState pulser
+        rsEnacted pulserRatifyState `shouldBe` Seq.empty
+        enactState <- getEnactState
+        rsEnactState pulserRatifyState `shouldBe` enactState
